Promote fResultLength field of TResult to 32 bits variable to avoid alignment issues on architectures where alignment matters. In case stdint wouldn't be available, default it to 'unsigned int'.
Fix Pi example accordingly. This break compatibility with previous results

+++ examples/Pi/result.c
@@ -44,7 +44,7 @@ int main(int argc, char *argv[])
         }
 
         /* Sanity check: verify we have correct length */
-        if (fread(&len, sizeof(unsigned short), 1, inFD) != 1)
+        if (fread(&len, sizeof(uint32_t), 1, inFD) != 1)
         {
             break;
         }
